add search across all products

+++ shop/shop_cite/views/cart_page.py
@@ -16,33 +16,37 @@ class CartPage(BaseTemplate):
                                context={'basket_items': basket_items})
 
     def post(self, request):
-        basket = request.session.get('basket', None)
-        for key in request.POST.keys():
-            if 'delete_item' in key:
-                product_id = key.split('-')[1]
-                if product_id in basket:
-                    del basket[product_id]
-            if 'remove_item' in key:
-                product_id = key.split('-')[1]
-                if product_id in basket:
-                    cnt = basket[product_id]['amount']
-                    if cnt > 1:
-                        basket[product_id]['amount'] = cnt - 1
-                    else:
+        srch_page = super().post(request)
+        if srch_page is None:
+            basket = request.session.get('basket', None)
+            for key in request.POST.keys():
+                if 'delete_item' in key:
+                    product_id = key.split('-')[1]
+                    if product_id in basket:
                         del basket[product_id]
-            if 'add_item' in key:
-                product_id = key.split('-')[1]
-                if product_id in basket:
-                    cnt = basket[product_id]['amount']
-                    basket[product_id]['amount'] = cnt + 1
+                if 'remove_item' in key:
+                    product_id = key.split('-')[1]
+                    if product_id in basket:
+                        cnt = basket[product_id]['amount']
+                        if cnt > 1:
+                            basket[product_id]['amount'] = cnt - 1
+                        else:
+                            del basket[product_id]
+                if 'add_item' in key:
+                    product_id = key.split('-')[1]
+                    if product_id in basket:
+                        cnt = basket[product_id]['amount']
+                        basket[product_id]['amount'] = cnt + 1
 
-        request.session['basket'] = basket
-        basket_items = []
-        if basket is not None:
-            for product in basket:
-                data = basket[product]
-                data['id'] = product
-                basket_items.append(data)
-        return self.get_render(request,
-                               'shop_cite/cart.html',
-                               context={'basket_items': basket_items})
+            request.session['basket'] = basket
+            basket_items = []
+            if basket is not None:
+                for product in basket:
+                    data = basket[product]
+                    data['id'] = product
+                    basket_items.append(data)
+            return self.get_render(request,
+                                   'shop_cite/cart.html',
+                                   context={'basket_items': basket_items})
+        return srch_page
+
